Move utilitary method to tools
