+++ scilpy/segment/tractogram_from_roi.py
@@ -7,8 +7,9 @@ import logging
 import nibabel as nib
 import numpy as np
 import os
+
+from scilpy.tractanalysis.tools import split_heads_tails_kmeans
 from scipy.ndimage import binary_dilation
-from sklearn.cluster import KMeans
 
 from dipy.io.stateful_tractogram import StatefulTractogram
 from dipy.io.streamline import save_tractogram
@@ -110,35 +111,7 @@ def compute_masks(gt_files, parser, args):
     return gt_bundle_masks, gt_bundle_inv_masks, affine, dimensions
 
 
-def split_heads_tails_kmeans(data):
-    """
-    Split a mask between head and tail with k means.
-
-    Parameters
-    ----------
-    data: numpy.ndarray
-        Mask to be split.
-
-    Returns
-    -------
-    mask_1: numpy.ndarray
-        "Head" of the mask.
-    mask_2: numpy.ndarray
-        "Tail" of the mask.
-    """
-
-    X = np.argwhere(data)
-    k_means = KMeans(n_clusters=2).fit(X)
-    mask_1 = np.zeros(data.shape)
-    mask_2 = np.zeros(data.shape)
-
-    mask_1[tuple(X[np.where(k_means.labels_ == 0)].T)] = 1
-    mask_2[tuple(X[np.where(k_means.labels_ == 1)].T)] = 1
-
-    return mask_1, mask_2
-
-
-def extract_tails_heads_from_endpoints(gt_endpoints, out_dir):
+def extract_and_save_tails_heads_from_endpoints(gt_endpoints, out_dir):
     """
     Extract two masks from a single mask containing two regions.
 
@@ -209,7 +182,7 @@ def compute_endpoint_masks(roi_options, affine, dimensions, out_dir):
     for bundle_options in roi_options:
         if 'gt_endpoints' in bundle_options:
             tail, head, _affine, _dimensions = \
-                extract_tails_heads_from_endpoints(
+                extract_and_save_tails_heads_from_endpoints(
                     bundle_options['gt_endpoints'], out_dir)
             if affine is not None:
                 # Compare affine
